pdf_compare_app/requirements.txt
streamlit
PyMuPDF
opencv-python
numpy
scikit-image
Pillow